This commit was manufactured by cvs2svn to create tag 'pre_5_0_dev2_m20041107_1'.
git-svn-id: https://svn.code.sf.net/p/seq/svn/showeq/tags/pre_5_0_dev2_m20041107_1@544 19a12193-7728-0410-8f54-e9f61ebc0c9f

+++ src/player.h
@@ -132,7 +132,7 @@ public:
  signals:
    void newPlayer(void);
    void buffLoad(const spellBuff*); 
-   void newSpeed               (int speed);
+   void newSpeed               (double speed);
    void statChanged            ( int statNum,
                                  int val,
                                  int max
